move functions to Analyses APIs
